moved ProfileCharts to single directory
